数据迁移

+++ mysql-oracle/src/main/java/com/example/mysqloracle/entity/news/LifeProductResult.java
@@ -30,7 +30,9 @@ public class LifeProductResult implements Serializable {
     @JsonProperty("type")
     private Integer type;
     @JsonProperty("company_id")
-    private Integer companyId;
+    private Long companyId;
+//    @JsonProperty("company_id")
+//    private Integer companyId;
     @JsonProperty("code")
     private String code;
     @JsonProperty("code_api")
